+++ src/blocks/notmuch.rs
@@ -13,21 +13,19 @@
 //!
 //! Key | Values | Default
 //! ----|--------|--------
-//! `format` | A string to customise the output of this block. See below for available placeholders. | `" $icon{ $name|} $count "`
+//! `format` | A string to customise the output of this block. See below for available placeholders. | `" $icon $count "`
 //! `maildir` | Path to the directory containing the notmuch database. Supports path expansions e.g. `~`. | `~/.mail`
 //! `query` | Query to run on the database. | `""`
 //! `threshold_critical` | Mail count that triggers `critical` state. | `99999`
 //! `threshold_warning` | Mail count that triggers `warning` state. | `99999`
 //! `threshold_good` | Mail count that triggers `good` state. | `99999`
 //! `threshold_info` | Mail count that triggers `info` state. | `99999`
-//! `name` | Label to show before the mail count. | `None`
 //! `interval` | Update interval in seconds. | `10`
 //!
 //! Placeholder | Value                                      | Type   | Unit
 //! ------------|--------------------------------------------|--------|-----
 //! `icon`      | A static icon                              | Icon   | -
 //! `count`     | Number of messages for the query           | Number | -
-//! `name`      | Value from config (if set)                 | Text   | -
 //!
 //! # Example
 //!
@@ -37,7 +35,6 @@
 //! query = "tag:alert and not tag:trash"
 //! threshold_warning = 1
 //! threshold_critical = 10
-//! name = "A"
 //! ```
 //!
 //! # Icons Used
@@ -62,13 +59,12 @@ struct NotmuchConfig {
     threshold_info: u32,
     #[default(u32::MAX)]
     threshold_good: u32,
-    name: Option<String>,
 }
 
 pub async fn run(config: toml::Value, mut api: CommonApi) -> Result<()> {
     let config = NotmuchConfig::deserialize(config).config_error()?;
     let mut widget = api.new_widget().with_format(
-        config.format.with_default(" $icon{ $name|} $count ")?,
+        config.format.with_default(" $icon $count ")?,
     );
 
     let db = config.maildir.expand()?;
@@ -80,8 +76,7 @@ pub async fn run(config: toml::Value, mut api: CommonApi) -> Result<()> {
 
         widget.set_values(map! {
             "icon" => Value::icon(api.get_icon("mail")?),
-            "count" => Value::number(count),
-            [if let Some(name) = &config.name] "name" => Value::text(name.to_string())
+            "count" => Value::number(count)
         });
 
         widget.state = if count >= config.threshold_critical {